@@ -4,10 +4,12 @@
   "description": "Theme boilerplate for WordPress using WP REST API and React.js",
   "private": true,
   "scripts": {
-    "start": "node run start",
-    "build": "node run build --production",
+    "start": "node scripts/start.js",
+    "build": "node scripts/build.js",
+    "old_start": "node run start",
+    "old_build": "node run build --production",
     "vccw:clone": "git clone https://github.com/vccw-team/vccw",
-    "vccw:checkout": "cd vccw && git checkout refs/tags/3.0.5",
+    "vccw:checkout": "cd vccw && git checkout refs/tags/3.18.0",
     "vccw:copy": "cp -r ./vccw/provision ./provision && cp ./vccw/Vagrantfile ./Vagrantfile && cp ./vccw/ansible.cfg ./ansible.cfg",
     "vccw:remove": "rm -rf ./vccw",
     "vccw": "yarn run vccw:clone && yarn run vccw:checkout && yarn run vccw:copy && yarn run vccw:remove",
@@ -15,7 +17,6 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "babel-core": "^6.0.20",
     "babel-polyfill": "^6.16.0",
     "babel-preset-es2015": "^6.0.15",
     "babel-preset-react": "^6.0.15",
@@ -26,37 +27,61 @@
     "react-redux": "^5.0.1",
     "react-router": "^3.0.0",
     "react-router-redux": "^4.0.7",
+    "react-scripts": "^1.1.4",
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0",
     "wpapi": "^1.0.0"
   },
   "devDependencies": {
-    "babel-loader": "^6.0.1",
+    "autoprefixer": "^7.1.6",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-runtime": "^6.26.0",
     "browser-sync": "^2.18.2",
     "bs-html-injector": "^3.0.3",
+    "case-sensitive-paths-webpack-plugin": "^2.1.1",
     "chokidar": "^1.6.1",
     "copy": "^0.3.0",
+    "css-loader": "^0.28.7",
     "del": "^2.2.2",
+    "eslint": "^4.19.1",
+    "eslint-config-react-app": "^2.1.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-flowtype": "^2.50.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-node": "^7.0.1",
+    "eslint-plugin-promise": "^3.8.0",
+    "eslint-plugin-react": "^7.10.0",
+    "eslint-plugin-standard": "^3.1.0",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^2.29.0",
     "json-loader": "^0.5.4",
+    "postcss-loader": "^2.0.8",
+    "react-dev-utils": "^5.0.1",
     "react-hot-loader": "^3.0.0-beta.6",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.1.1",
     "standard": "^8.3.0",
     "standard-loader": "^5.0.0",
-    "webpack": "^1.13.3",
+    "style-loader": "^0.19.0",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.8.1",
     "webpack-dev-middleware": "^1.8.4",
+    "webpack-dev-server": "^2.9.4",
     "webpack-hot-middleware": "^2.13.2",
     "write-file-webpack-plugin": "^3.4.2"
   },
   "babel": {
     "presets": [
-      "es2015",
-      "stage-0",
-      "react"
+      "react-app"
     ],
     "plugins": [
       "react-hot-loader/babel"
     ]
+  },
+  "eslintConfig": {
+    "extends": "react-app"
   }
 }
