@@ -1,12 +1,13 @@
-import React from 'react'
-import { Router, Route, IndexRoute } from 'react-router'
-import ReactGA from 'react-ga'
-import Root from './containers/Root'
-import Frontpage from './containers/Frontpage'
-import Archive from './containers/Archive'
-import Single from './containers/Single'
-import Page from './containers/Page'
-import NoMatch from './containers/NoMatch'
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { Router, Route, IndexRoute } from 'react-router';
+import ReactGA from 'react-ga';
+import Root from './containers/Root';
+import Frontpage from './containers/Frontpage';
+import Archive from './containers/Archive';
+import Single from './containers/Single';
+import Page from './containers/Page';
+import NoMatch from './containers/NoMatch';
 
 ReactGA.initialize('UA-000000-01')
 
