@@ -1,12 +1,12 @@
-const WP_PARAMS = global.WP_PARAMETERS
+const WP_PARAMS = global.WP_PARAMETERS;
 
-export const WP_URL = WP_PARAMS.API
-export const WP_SITE_TITLE = WP_PARAMS.SITE_TITLE
-export const WP_SITE_DESCRIPTION = WP_PARAMS.SITE_DESCRIPTION
-export const WP_API = WP_PARAMS.API
-export const WP_PAGE_ON_FRONT = WP_PARAMS.PAGE_ON_FRONT
-export const WP_POSTS_PER_PAGE = WP_PARAMS.POSTS_PER_PAGE
+export const WP_URL = WP_PARAMS.API;
+export const WP_SITE_TITLE = WP_PARAMS.SITE_TITLE;
+export const WP_SITE_DESCRIPTION = WP_PARAMS.SITE_DESCRIPTION;
+export const WP_API = WP_PARAMS.API;
+export const WP_PAGE_ON_FRONT = WP_PARAMS.PAGE_ON_FRONT;
+export const WP_POSTS_PER_PAGE = WP_PARAMS.POSTS_PER_PAGE;
 
-export const REQUEST_API = 'REQUEST_API'
-export const RECEIVE_ARCHIVES = 'RECEIVE_ARCHIVES'
-export const RECEIVE_PAGE = 'RECEIVE_PAGE'
+export const REQUEST_API = "REQUEST_API";
+export const RECEIVE_ARCHIVES = "RECEIVE_ARCHIVES";
+export const RECEIVE_PAGE = "RECEIVE_PAGE";
